Consolidate modules and update import
