+++ test/test_functions.py
@@ -1,0 +1,8 @@
+from src.functions import combine
+
+
+def test_combine():
+    expected = "hello world"
+    actual = combine("hello ", "world")
+
+    assert expected == actual
